requirements.txt
# Core RAG Frameworks
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
llama-index>=0.10.0

# Vector Stores
chromadb>=0.4.22
faiss-cpu>=1.7.4
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.3.0
openai>=1.10.0

# Document Processing
pymupdf>=1.23.0
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0

# Audio Processing
openai-whisper>=20231117
pydub>=0.25.1

# Image Processing
pillow>=10.2.0
pytesseract>=0.3.10

# Knowledge Graph
networkx>=3.2.0
neo4j>=5.15.0
pyvis>=0.3.2

# Search & Retrieval
rank-bm25>=0.2.2

# API & Web
requests>=2.31.0
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Progress & UI
tqdm>=4.66.0
rich>=13.7.0

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0

# Development
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
pytest>=8.0.0
mypy>=1.8.0

